refactor(image-analysis): tighten config/ file boundaries
Two file-boundary smells exposed once the alias registry and the
``image_kind`` / ``scan_type`` discriminators were both gone:

1. ``aliases.py`` had no aliases anymore — just the
   ``ImageAnalyzerSpec`` field model + its surface-form parser. A
   helper file for ``diagnostic.py`` pretending to be its own concern.
2. ``factory.py`` mixed two concerns: ``load_diagnostic`` (file →
   typed config) and ``create_image_analyzer`` (typed config → live
   instance). The "factory" file was half loader.

New layout — one concern per file, names match contents:

* ``diagnostic.py``: all diagnostic-level models —
  ``DiagnosticAnalysisConfig``, ``ImageAnalyzerSpec``,
  ``resolve_image_analyzer_value``, ``ImageSection``. Absorbs the
  contents of the deleted ``aliases.py``.
* ``loader.py``: file → typed config. Absorbs ``load_diagnostic``
  (plus its private helpers ``_resolve_default_config_dir`` and
  ``_discover_analyzers``) from ``factory.py``. Now hosts
  ``load_camera_config`` / ``load_line_config`` /
  ``load_diagnostic`` / ``find_config_file`` — every "turn a file
  into a typed config" entry point lives here.
* ``factory.py``: typed config → live instance. Just
  ``create_image_analyzer`` and its private helpers.

No behavior change; no public-API change. The package ``__init__``
re-exports the same names from their new homes, so external callers
that imported from ``image_analysis.config`` are unaffected.
