+++ ImageAnalysis/image_analysis/config/diagnostic.py
@@ -1,14 +1,14 @@
-"""Top-level Pydantic model for the unified diagnostic config schema.
+"""Diagnostic-level Pydantic models for the unified diagnostic config schema.
 
 A diagnostic config is one YAML file per device that bundles two
 concerns previously split across ``image_analysis_configs/`` and
 ``scan_analysis_configs/library/analyzers/``:
 
 * ``image:`` — the ImageAnalysis-owned section. A typed
-  :class:`CameraConfig` (``type: camera``, the default) or
-  :class:`Line1DConfig` (``type: line``). Pydantic's discriminated
-  union routes the dict to the right model based on the ``type``
-  field. Attribute access on ``diag.image`` works directly
+  :class:`CameraConfig` (``type: camera``) or :class:`Line1DConfig`
+  (``type: line``). Pydantic's discriminated union routes the dict
+  to the right model based on the ``type`` field. Attribute access
+  on ``diag.image`` works directly
   (``diag.image.background.method = "constant"``).
 * ``scan:`` — the ScanAnalysis-owned section (priority, mode, save
   flags, gdoc slot, file tail, renderer kwargs). Kept weakly typed
@@ -19,18 +19,17 @@ concerns previously split across ``image_analysis_configs/`` and
 
 A diagnostic also declares its ImageAnalyzer via ``image_analyzer``,
 either as a bare class-path string or as a verbose ``{class_path,
-kwargs}`` dict. Both forms normalise to :class:`ImageAnalyzerSpec`.
+kwargs}`` dict — both forms validate into :class:`ImageAnalyzerSpec`.
 
 The 2D-vs-1D dimension lives in one place: the ``type`` field on the
-image section. Both ``image_analyzer.image_kind`` and
-``image_analyzer.scan_type`` are gone — they were sibling
-discriminators describing the shape of ``image:``, and now the
-``image:`` section discriminates itself.
+image section. The ScanAnalyzer wrapper (Array2D vs Array1D) is
+picked the same way at scan-build time, from the type of
+``diag.image``.
 """
 
 from __future__ import annotations
 
-from typing import Annotated, Any, Dict, Optional, Union
+from typing import Annotated, Any, Dict, Mapping, Optional, Union
 
 from pydantic import (
     BaseModel,
@@ -40,9 +39,114 @@ from pydantic import (
     model_validator,
 )
 
-from .aliases import ImageAnalyzerSpec, resolve_image_analyzer_value
 from .array1d_processing import Line1DConfig
 from .array2d_processing import CameraConfig
+
+
+# ---------------------------------------------------------------------------
+# image_analyzer field model + surface-form normalisation
+# ---------------------------------------------------------------------------
+
+
+class ImageAnalyzerSpec(BaseModel):
+    """Canonical runtime form of a diagnostic's ``image_analyzer`` field.
+
+    Whether the YAML used the bare-string form or the verbose dict,
+    the field validator on
+    :class:`DiagnosticAnalysisConfig.image_analyzer` normalises to
+    this shape before anything else sees it.
+
+    Two surface forms are accepted on the ``image_analyzer`` field:
+
+    * **Bare string** — the class path. Right for the common case
+      (``BeamAnalyzer``, ``StandardAnalyzer``, ``GrenouilleAnalyzer``,
+      ``MagSpecManualCalibAnalyzer``, …)::
+
+          image_analyzer: image_analysis.analyzers.beam_analyzer.BeamAnalyzer
+
+    * **Verbose dict** — class path plus extra constructor kwargs.
+      Right for analyzers that take per-instance kwargs (HASO's
+      ``wavekit_config_file_path``, for example)::
+
+          image_analyzer:
+            class_path: image_analysis.analyzers.HASO_himg_has_processor.HASOHimgHasProcessor
+            kwargs:
+              wavekit_config_file_path: /path/to/wfs.dat
+              mask_top: 125
+
+    Attributes
+    ----------
+    class_path : str
+        Fully qualified import path of the ImageAnalyzer class
+        (``"image_analysis.analyzers.beam_analyzer.BeamAnalyzer"``).
+    kwargs : dict
+        Per-instance kwargs passed to the analyzer's constructor. The
+        loader injects the resolved embedded image config under the
+        constructor's expected name (``camera_config`` or
+        ``line_config``) separately; do not put the image config here.
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    class_path: str = Field(min_length=1)
+    kwargs: Dict[str, Any] = Field(default_factory=dict)
+
+
+def resolve_image_analyzer_value(value: Any) -> Dict[str, Any]:
+    """Normalise an ``image_analyzer`` YAML value to an :class:`ImageAnalyzerSpec` dict.
+
+    Accepts the two surface forms documented on :class:`ImageAnalyzerSpec`
+    and returns a dict ready to be validated. Used by the field
+    validator on :class:`DiagnosticAnalysisConfig.image_analyzer`.
+
+    Parameters
+    ----------
+    value : Any
+        Whatever the YAML deserialiser produced for the
+        ``image_analyzer`` field — a bare class-path string or a dict
+        with ``class_path`` (or YAML-friendly ``class``) plus optional
+        ``kwargs``.
+
+    Returns
+    -------
+    dict
+        A dict matching :class:`ImageAnalyzerSpec`'s field layout.
+
+    Raises
+    ------
+    ValueError
+        If the dict is missing ``class_path`` / ``class``, or if an
+        unrecognised input shape is supplied.
+    """
+    if isinstance(value, str):
+        return {"class_path": value}
+
+    if isinstance(value, Mapping):
+        data = dict(value)
+        if "class" in data:
+            # YAML-friendly alias for the model field.
+            data["class_path"] = data.pop("class")
+        if "class_path" not in data:
+            raise ValueError(
+                "image_analyzer dict must contain 'class_path' (or "
+                "YAML-friendly 'class') naming the analyzer class. "
+                f"Got keys: {sorted(data)}"
+            )
+        return data
+
+    # Already an ImageAnalyzerSpec instance — let Pydantic pass it through.
+    if isinstance(value, ImageAnalyzerSpec):
+        return value.model_dump()
+
+    raise ValueError(
+        f"image_analyzer must be a class-path string or a mapping; got "
+        f"{type(value).__name__}"
+    )
+
+
+# ---------------------------------------------------------------------------
+# image: section — discriminated union of camera / line configs
+# ---------------------------------------------------------------------------
 
 
 # Pydantic uses the ``type`` field on each variant
@@ -56,6 +160,11 @@ ImageSection = Annotated[
     Union[CameraConfig, Line1DConfig],
     Field(discriminator="type"),
 ]
+
+
+# ---------------------------------------------------------------------------
+# Top-level diagnostic model
+# ---------------------------------------------------------------------------
 
 
 class DiagnosticAnalysisConfig(BaseModel):
